Changing handling of Request Headers and Curl Opts
Upon initialization, two sets of curl opt and request header arrays are created.  They both contain the same values initially, however one is editable and the other is not.  An end user may modify the contents of these two arrays utilizing the appropriate public method without interference by SoapClientPlus.  If the user wishes to reset to default values, they may do so with the use of the appropriate "reset" method